requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# AI and Communication
openai==1.3.0
twilio==8.10.0

# Redis for core state management
redis==5.0.1
celery==5.3.4

# Object storage (AWS S3 compatible)
boto3==1.34.0
botocore==1.34.0

# Phone number utilities
phonenumbers==8.13.26

# Observability and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Background task processing
asyncio-mqtt==0.16.1

# Resume parsing dependencies
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
